requirements.txt
pyzmq
nose
sphinx
inflection

elasticsearch